+++ test/codecTests.hs
@@ -165,7 +165,7 @@ tests =
           Right _ -> assertFailure "Builder should have failed with duplicate beans error",
       testCase "should fail cycle wiring" do
         Data.Foldable.for_ @[] [("forbid", forbidDepCycles), ("selfdeps", allowSelfDeps)] \(name, fire) ->
-          case cook @(Serializer Foo) fire [cauldron] of
+          case cook @(Serializer Foo) fire cauldron of
             Left (DependencyCycleError _) -> pure ()
             Left _ -> assertFailure $ "Unexpected error when wiring" ++ name
             Right _ -> assertFailure $ "Unexpected success when wiring" ++ name,
@@ -175,7 +175,7 @@ tests =
             ("someConsume", cauldronAccums2, Acc 10)
           ]
           \(name, c, expected) ->
-            case cook @Acc allowDepCycles [c] of
+            case cook @Acc allowDepCycles c of
               Left _err -> do
                 -- putStrLn $ prettyRecipeError err
                 assertFailure $ "could not wire " ++ name
@@ -187,7 +187,7 @@ tests =
             ("indirectagg", cauldronAccumsOops2)
           ]
           \(name, c) ->
-            case cook @(Serializer Foo) allowDepCycles [c] of
+            case cook @(Serializer Foo) allowDepCycles c of
               Left (DependencyCycleError _) -> assertFailure $ "We should be able to wire cycles with accs"
               Left _ -> assertFailure $ "Unexpected error when wiring" ++ name
               Right _ -> pure ()
@@ -195,7 +195,7 @@ tests =
   where
     makeBasicTest :: Cauldron Identity -> IO ()
     makeBasicTest theCauldron =
-      case cook allowDepCycles [theCauldron] of
+      case cook allowDepCycles theCauldron of
         Left _ -> do
           -- putStrLn $ prettyRecipeError err
           assertFailure "could not wire"